+++ src/main/java/com/braintreegateway/MerchantAccount.java
@@ -12,9 +12,30 @@ public class MerchantAccount {
         UNRECOGNIZED
     }
 
+    public enum FundingDestination {
+        BANK("bank"),
+        MOBILE_PHONE("mobile_phone"),
+        EMAIL("email"),
+        UNRECOGNIZED("unrecognized");
+
+        private final String name;
+
+        FundingDestination(String name) {
+            this.name = name;
+        }
+
+        @Override
+        public String toString() {
+            return name;
+        }
+    }
+
     private final String id;
     private final Status status;
     private final MerchantAccount masterMerchantAccount;
+    private final IndividualDetails individualDetails;
+    private final BusinessDetails businessDetails;
+    private final FundingDetails fundingDetails;
 
     public MerchantAccount(NodeWrapper node) {
         this.id = node.findString("id");
@@ -25,6 +46,24 @@ public class MerchantAccount {
             this.masterMerchantAccount = new MerchantAccount(masterNode);
         else
             this.masterMerchantAccount = null;
+
+        NodeWrapper individualNode = node.findFirst("individual");
+        if (individualNode != null)
+            this.individualDetails = new IndividualDetails(individualNode);
+        else
+            this.individualDetails = null;
+
+        NodeWrapper businessNode = node.findFirst("business");
+        if (businessNode != null)
+            this.businessDetails = new BusinessDetails(businessNode);
+        else
+            this.businessDetails = null;
+
+        NodeWrapper fundingNode = node.findFirst("funding");
+        if (fundingNode != null)
+            this.fundingDetails = new FundingDetails(fundingNode);
+        else
+            this.fundingDetails = null;
     }
 
     public String getId() {
@@ -41,6 +80,18 @@ public class MerchantAccount {
      */
     public MerchantAccount getMasterMerchantAccount() {
         return masterMerchantAccount;
+    }
+
+    public IndividualDetails getIndividualDetails() {
+        return individualDetails;
+    }
+
+    public BusinessDetails getBusinessDetails() {
+        return businessDetails;
+    }
+
+    public FundingDetails getFundingDetails() {
+        return fundingDetails;
     }
 
     public boolean isSubMerchant() {