+++ src/main/java/com/braintreegateway/MerchantAccountGateway.java
@@ -17,4 +17,9 @@ public final class MerchantAccountGateway {
         final NodeWrapper response = http.post(CREATE_URL, request);
         return new Result<MerchantAccount>(response, MerchantAccount.class);
     }
+
+    public Result<MerchantAccount> update(String id, MerchantAccountRequest request) {
+        final NodeWrapper response = http.put("/merchant_accounts/" + id + "/update_via_api", request);
+        return new Result<MerchantAccount>(response, MerchantAccount.class);
+    }
 }