+++ src/main/java/com/braintreegateway/MerchantAccountRequest.java
@@ -6,7 +6,10 @@ package com.braintreegateway;
 public class MerchantAccountRequest extends Request {
 
     private ApplicantDetailsRequest applicantDetails;
-    private boolean tosAccepted;
+    private IndividualRequest individual;
+    private BusinessRequest business;
+    private FundingRequest funding;
+    private Boolean tosAccepted;
     private String masterMerchantAccountId;
     private String id;
 
@@ -18,6 +21,21 @@ public class MerchantAccountRequest extends Request {
     public ApplicantDetailsRequest applicantDetails() {
         applicantDetails = new ApplicantDetailsRequest(this);
         return applicantDetails;
+    }
+
+    public IndividualRequest individual() {
+        individual = new IndividualRequest(this);
+        return individual;
+    }
+
+    public BusinessRequest business() {
+        business = new BusinessRequest(this);
+        return business;
+    }
+
+    public FundingRequest funding() {
+        funding = new FundingRequest(this);
+        return funding;
     }
 
     public MerchantAccountRequest masterMerchantAccountId(String masterMerchantAccountId) {
@@ -48,6 +66,9 @@ public class MerchantAccountRequest extends Request {
     protected RequestBuilder buildRequest(String root) {
         return new RequestBuilder(root).
                 addElement("applicantDetails", applicantDetails).
+                addElement("individual", individual).
+                addElement("business", business).
+                addElement("funding", funding).
                 addElement("tosAccepted", tosAccepted).
                 addElement("masterMerchantAccountId", masterMerchantAccountId).
                 addElement("id", id);